@@ -3,12 +3,13 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.ML.Runtime.CommandLine;
+using Microsoft.ML.Runtime.EntryPoints;
 using Microsoft.ML.Runtime.Internal.Utilities;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using System.Text.RegularExpressions;
 
 // REVIEW: Determine ideal namespace.
 namespace Microsoft.ML.Runtime
@@ -26,9 +27,14 @@ namespace Microsoft.ML.Runtime
         {
             _lock = new object();
             _cachedAssemblies = new HashSet<string>();
-            _classesByKey = new ConcurrentDictionary<LoadableClassInfo.Key, LoadableClassInfo>();
-            _classes = new ConcurrentQueue<LoadableClassInfo>();
-            _signatures = new ConcurrentDictionary<Type, bool>();
+            _classesByKey = new Dictionary<LoadableClassInfo.Key, LoadableClassInfo>();
+            _classes = new List<LoadableClassInfo>();
+            _signatures = new Dictionary<Type, bool>();
+
+            _entryPoints = new List<EntryPointInfo>();
+            _entryPointMap = new Dictionary<string, EntryPointInfo>();
+            _componentMap = new Dictionary<string, ComponentInfo>();
+            _components = new List<ComponentInfo>();
         }
 
         /// <summary>
@@ -255,18 +261,138 @@ namespace Microsoft.ML.Runtime
             }
         }
 
-        // This lock protects _cachedAssemblies only. The collection of ClassInfos is concurrent
-        // so needs no protection.
+        /// <summary>
+        /// A description of a single entry point.
+        /// </summary>
+        public sealed class EntryPointInfo
+        {
+            public readonly string Name;
+            public readonly string Description;
+            public readonly string ShortName;
+            public readonly string FriendlyName;
+            public readonly string[] XmlInclude;
+            public readonly MethodInfo Method;
+            public readonly Type InputType;
+            public readonly Type OutputType;
+            public readonly Type[] InputKinds;
+            public readonly Type[] OutputKinds;
+            public readonly ObsoleteAttribute ObsoleteAttribute;
+
+            internal EntryPointInfo(MethodInfo method,
+                TlcModule.EntryPointAttribute attribute, ObsoleteAttribute obsoleteAttribute)
+            {
+                Contracts.AssertValue(method);
+                Contracts.AssertValue(attribute);
+
+                Name = attribute.Name ?? string.Join(".", method.DeclaringType.Name, method.Name);
+                Description = attribute.Desc;
+                Method = method;
+                ShortName = attribute.ShortName;
+                FriendlyName = attribute.UserName;
+                XmlInclude = attribute.XmlInclude;
+                ObsoleteAttribute = obsoleteAttribute;
+
+                // There are supposed to be 2 parameters, env and input for non-macro nodes.
+                // Macro nodes have a 3rd parameter, the entry point node.
+                var parameters = method.GetParameters();
+                if (parameters.Length != 2 && parameters.Length != 3)
+                    throw Contracts.Except("Method '{0}' has {1} parameters, but must have 2 or 3", method.Name, parameters.Length);
+                if (parameters[0].ParameterType != typeof(IHostEnvironment))
+                    throw Contracts.Except("Method '{0}', 1st parameter is {1}, but must be IHostEnvironment", method.Name, parameters[0].ParameterType);
+                InputType = parameters[1].ParameterType;
+                var outputType = method.ReturnType;
+                if (!outputType.IsClass)
+                    throw Contracts.Except("Method '{0}' returns {1}, but must return a class", method.Name, outputType);
+                OutputType = outputType;
+
+                InputKinds = FindEntryPointKinds(InputType);
+                OutputKinds = FindEntryPointKinds(OutputType);
+            }
+
+            private Type[] FindEntryPointKinds(Type type)
+            {
+                var kindAttr = type.GetTypeInfo().GetCustomAttributes(typeof(TlcModule.EntryPointKindAttribute), false).FirstOrDefault()
+                    as TlcModule.EntryPointKindAttribute;
+                var baseType = type.BaseType;
+
+                if (baseType == null)
+                    return kindAttr?.Kinds;
+                var baseKinds = FindEntryPointKinds(baseType);
+                if (kindAttr == null)
+                    return baseKinds;
+                if (baseKinds == null)
+                    return kindAttr.Kinds;
+                return kindAttr.Kinds.Concat(baseKinds).ToArray();
+            }
+
+            public override string ToString() => $"{Name}: {Description}";
+        }
+
+        /// <summary>
+        /// A description of a single component.
+        /// The 'component' is a non-standalone building block that is used to parametrize entry points or other ML.NET components.
+        /// For example, 'Loss function', or 'similarity calculator' could be components.
+        /// </summary>
+        public sealed class ComponentInfo
+        {
+            public readonly string Name;
+            public readonly string Description;
+            public readonly string FriendlyName;
+            public readonly string Kind;
+            public readonly Type ArgumentType;
+            public readonly Type InterfaceType;
+            public readonly string[] Aliases;
+
+            internal ComponentInfo(Type interfaceType, string kind, Type argumentType, TlcModule.ComponentAttribute attribute)
+            {
+                Contracts.AssertValue(interfaceType);
+                Contracts.AssertNonEmpty(kind);
+                Contracts.AssertValue(argumentType);
+                Contracts.AssertValue(attribute);
+
+                Name = attribute.Name;
+                Description = attribute.Desc;
+                if (string.IsNullOrWhiteSpace(attribute.FriendlyName))
+                    FriendlyName = Name;
+                else
+                    FriendlyName = attribute.FriendlyName;
+
+                Kind = kind;
+                if (!IsValidName(Kind))
+                    throw Contracts.Except("Invalid component kind: '{0}'", Kind);
+
+                Aliases = attribute.Aliases;
+                if (!IsValidName(Name))
+                    throw Contracts.Except("Component name '{0}' is not valid.", Name);
+
+                if (Aliases != null && Aliases.Any(x => !IsValidName(x)))
+                    throw Contracts.Except("Component '{0}' has an invalid alias '{1}'", Name, Aliases.First(x => !IsValidName(x)));
+
+                if (!typeof(IComponentFactory).IsAssignableFrom(argumentType))
+                    throw Contracts.Except("Component '{0}' must inherit from IComponentFactory", argumentType);
+
+                ArgumentType = argumentType;
+                InterfaceType = interfaceType;
+            }
+        }
+
+        // This lock protects adding to the below collections.
         private readonly object _lock;
         private readonly HashSet<string> _cachedAssemblies;
 
         // Map from key/name to loadable class. Note that the same ClassInfo may appear
-        // multiple times. For the set of unique infos, use s_classes.
-        private readonly ConcurrentDictionary<LoadableClassInfo.Key, LoadableClassInfo> _classesByKey;
+        // multiple times. For the set of unique infos, use _classes.
+        private readonly Dictionary<LoadableClassInfo.Key, LoadableClassInfo> _classesByKey;
 
         // The unique ClassInfos and Signatures.
-        private readonly ConcurrentQueue<LoadableClassInfo> _classes;
-        private readonly ConcurrentDictionary<Type, bool> _signatures;
+        private readonly List<LoadableClassInfo> _classes;
+        private readonly Dictionary<Type, bool> _signatures;
+
+        private readonly List<EntryPointInfo> _entryPoints;
+        private readonly Dictionary<string, EntryPointInfo> _entryPointMap;
+
+        private readonly List<ComponentInfo> _components;
+        private readonly Dictionary<string, ComponentInfo> _componentMap;
 
         private static bool TryGetIniters(Type instType, Type loaderType, Type[] parmTypes,
             out MethodInfo getter, out ConstructorInfo ctor, out MethodInfo create, out bool requireEnvironment)
@@ -296,28 +422,122 @@ namespace Microsoft.ML.Runtime
             return false;
         }
 
-        private void AddClass(LoadableClassInfo info, string[] loadNames)
+        private void AddClass(LoadableClassInfo info, string[] loadNames, bool throwOnError)
         {
-            _classes.Enqueue(info);
+            _classes.Add(info);
+            bool isEntryPoint = false;
             foreach (var sigType in info.SignatureTypes)
             {
-                _signatures.TryAdd(sigType, true);
+                _signatures[sigType] = true;
 
                 foreach (var name in loadNames)
                 {
                     string nameCi = name.ToLowerInvariant();
 
                     var key = new LoadableClassInfo.Key(nameCi, sigType);
-                    if (!_classesByKey.TryAdd(key, info))
+                    if (_classesByKey.TryGetValue(key, out var infoCur))
                     {
-                        var infoCur = _classesByKey[key];
-                        // REVIEW: Fix this message to reflect the signature....
-                        Console.Error.WriteLine(
-                            "CacheClassesFromAssembly: can't map name {0} to {1}, already mapped to {2}",
-                            name, info.Type.Name, infoCur.Type.Name);
+                        if (throwOnError)
+                        {
+                            throw Contracts.Except($"ComponentCatalog cannot map name '{name}' and SignatureType '{sigType}' to {info.Type.Name}, already mapped to {infoCur.Type.Name}.");
+                        }
+                    }
+                    else
+                    {
+                        _classesByKey.Add(key, info);
                     }
                 }
+
+                if (sigType == typeof(SignatureEntryPointModule))
+                {
+                    isEntryPoint = true;
+                }
             }
+
+            if (isEntryPoint)
+            {
+                ScanForEntryPoints(info);
+            }
+        }
+
+        private void ScanForEntryPoints(LoadableClassInfo info)
+        {
+            var type = info.LoaderType;
+
+            // Scan for entry points.
+            foreach (var methodInfo in type.GetMethods(BindingFlags.Static | BindingFlags.Public))
+            {
+                var attr = methodInfo.GetCustomAttributes(typeof(TlcModule.EntryPointAttribute), false).FirstOrDefault() as TlcModule.EntryPointAttribute;
+                if (attr == null)
+                    continue;
+
+                var entryPointInfo = new EntryPointInfo(methodInfo, attr,
+                    methodInfo.GetCustomAttributes(typeof(ObsoleteAttribute), false).FirstOrDefault() as ObsoleteAttribute);
+
+                _entryPoints.Add(entryPointInfo);
+                if (_entryPointMap.ContainsKey(entryPointInfo.Name))
+                {
+                    // Duplicate entry point name. We need to show a warning here.
+                    // REVIEW: we will be able to do this once catalog becomes a part of env.
+                    continue;
+                }
+
+                _entryPointMap[entryPointInfo.Name] = entryPointInfo;
+            }
+
+            // Scan for components.
+            // First scan ourself, and then all nested types, for component info.
+            ScanForComponents(type);
+            foreach (var nestedType in type.GetTypeInfo().GetNestedTypes())
+                ScanForComponents(nestedType);
+        }
+
+        private bool ScanForComponents(Type nestedType)
+        {
+            var attr = nestedType.GetTypeInfo().GetCustomAttributes(typeof(TlcModule.ComponentAttribute), true).FirstOrDefault()
+                as TlcModule.ComponentAttribute;
+            if (attr == null)
+                return false;
+
+            bool found = false;
+            foreach (var faceType in nestedType.GetInterfaces())
+            {
+                var faceAttr = faceType.GetTypeInfo().GetCustomAttributes(typeof(TlcModule.ComponentKindAttribute), false).FirstOrDefault()
+                        as TlcModule.ComponentKindAttribute;
+                if (faceAttr == null)
+                    continue;
+
+                if (!typeof(IComponentFactory).IsAssignableFrom(faceType))
+                    throw Contracts.Except("Component signature '{0}' doesn't inherit from '{1}'", faceType, typeof(IComponentFactory));
+
+                try
+                {
+                    // In order to populate from JSON, we need to invoke the parameterless ctor. Testing that this is possible.
+                    Activator.CreateInstance(nestedType);
+                }
+                catch (MissingMemberException ex)
+                {
+                    throw Contracts.Except(ex, "Component type '{0}' doesn't have a default constructor", faceType);
+                }
+
+                var info = new ComponentInfo(faceType, faceAttr.Kind, nestedType, attr);
+                var names = (info.Aliases ?? new string[0]).Concat(new[] { info.Name }).Distinct();
+                _components.Add(info);
+
+                foreach (var alias in names)
+                {
+                    var tag = $"{info.Kind}:{alias}";
+                    if (_componentMap.ContainsKey(tag))
+                    {
+                        // Duplicate component name. We need to show a warning here.
+                        // REVIEW: we will be able to do this once catalog becomes a part of env.
+                        continue;
+                    }
+                    _componentMap[tag] = info;
+                }
+            }
+
+            return found;
         }
 
         private static MethodInfo FindInstanceGetter(Type instType, Type loaderType)
@@ -379,15 +599,14 @@ namespace Microsoft.ML.Runtime
                         {
                             if (throwOnError)
                             {
-                                throw new InvalidOperationException(string.Format(
-                                    "Can't instantiate loadable class {0} with name {1}",
-                                    attr.InstanceType.Name, attr.LoadNames[0]));
+                                throw Contracts.Except(
+                                    $"Can't instantiate loadable class '{attr.InstanceType.Name}' with name '{attr.LoadNames[0]}'");
                             }
                             Contracts.Assert(getter == null && ctor == null && create == null);
                         }
                         var info = new LoadableClassInfo(attr, getter, ctor, create, requireEnvironment);
 
-                        AddClass(info, attr.LoadNames);
+                        AddClass(info, attr.LoadNames, throwOnError);
                     }
                 }
             }
@@ -488,6 +707,144 @@ namespace Microsoft.ML.Runtime
             Contracts.CheckValueOrNull(loadName);
             loadName = (loadName ?? "").ToLowerInvariant().Trim();
             return FindClassCore(new LoadableClassInfo.Key(loadName, signatureType));
+        }
+
+        /// <summary>
+        /// Get all registered entry points.
+        /// </summary>
+        public IEnumerable<EntryPointInfo> AllEntryPoints()
+        {
+            return _entryPoints.AsEnumerable();
+        }
+
+        public bool TryFindEntryPoint(string name, out EntryPointInfo entryPoint)
+        {
+            Contracts.CheckNonEmpty(name, nameof(name));
+            return _entryPointMap.TryGetValue(name, out entryPoint);
+        }
+
+        public bool TryFindComponent(string kind, string alias, out ComponentInfo component)
+        {
+            Contracts.CheckNonEmpty(kind, nameof(kind));
+            Contracts.CheckNonEmpty(alias, nameof(alias));
+
+            // Note that, if kind or alias contain the colon character, the kind:name 'tag' will contain more than one colon.
+            // Since colon may not appear in any valid name, the dictionary lookup is guaranteed to fail.
+            return _componentMap.TryGetValue($"{kind}:{alias}", out component);
+        }
+
+        public bool TryFindComponent(Type argumentType, out ComponentInfo component)
+        {
+            Contracts.CheckValue(argumentType, nameof(argumentType));
+
+            component = _components.FirstOrDefault(x => x.ArgumentType == argumentType);
+            return component != null;
+        }
+
+        public bool TryFindComponent(Type interfaceType, Type argumentType, out ComponentInfo component)
+        {
+            Contracts.CheckValue(interfaceType, nameof(interfaceType));
+            Contracts.CheckParam(interfaceType.IsInterface, nameof(interfaceType), "Must be interface");
+            Contracts.CheckValue(argumentType, nameof(argumentType));
+
+            component = _components.FirstOrDefault(x => x.InterfaceType == interfaceType && x.ArgumentType == argumentType);
+            return component != null;
+        }
+
+        public bool TryFindComponent(Type interfaceType, string alias, out ComponentInfo component)
+        {
+            Contracts.CheckValue(interfaceType, nameof(interfaceType));
+            Contracts.CheckParam(interfaceType.IsInterface, nameof(interfaceType), "Must be interface");
+            Contracts.CheckNonEmpty(alias, nameof(alias));
+            component = _components.FirstOrDefault(x => x.InterfaceType == interfaceType && (x.Name == alias || (x.Aliases != null && x.Aliases.Contains(alias))));
+            return component != null;
+        }
+
+        /// <summary>
+        /// Akin to <see cref="TryFindComponent(Type, string, out ComponentInfo)"/>, except if the regular (case sensitive) comparison fails, it will
+        /// attempt to back off to a case-insensitive comparison.
+        /// </summary>
+        public bool TryFindComponentCaseInsensitive(Type interfaceType, string alias, out ComponentInfo component)
+        {
+            Contracts.CheckValue(interfaceType, nameof(interfaceType));
+            Contracts.CheckParam(interfaceType.IsInterface, nameof(interfaceType), "Must be interface");
+            Contracts.CheckNonEmpty(alias, nameof(alias));
+            if (TryFindComponent(interfaceType, alias, out component))
+                return true;
+            alias = alias.ToLowerInvariant();
+            component = _components.FirstOrDefault(x => x.InterfaceType == interfaceType && (x.Name.ToLowerInvariant() == alias || AnyMatch(alias, x.Aliases)));
+            return component != null;
+        }
+
+        private static bool AnyMatch(string name, string[] aliases)
+        {
+            if (aliases == null)
+                return false;
+            return aliases.Any(a => string.Equals(name, a, StringComparison.OrdinalIgnoreCase));
+        }
+
+        /// <summary>
+        /// Returns all valid component kinds.
+        /// </summary>
+        public IEnumerable<string> GetAllComponentKinds()
+        {
+            return _components.Select(x => x.Kind).Distinct().OrderBy(x => x);
+        }
+
+        /// <summary>
+        /// Returns all components of the specified kind.
+        /// </summary>
+        public IEnumerable<ComponentInfo> GetAllComponents(string kind)
+        {
+            Contracts.CheckNonEmpty(kind, nameof(kind));
+            Contracts.CheckParam(IsValidName(kind), nameof(kind), "Invalid component kind");
+            return _components.Where(x => x.Kind == kind).OrderBy(x => x.Name);
+        }
+
+        /// <summary>
+        /// Returns all components that implement the specified interface.
+        /// </summary>
+        public IEnumerable<ComponentInfo> GetAllComponents(Type interfaceType)
+        {
+            Contracts.CheckValue(interfaceType, nameof(interfaceType));
+            return _components.Where(x => x.InterfaceType == interfaceType).OrderBy(x => x.Name);
+        }
+
+        public bool TryGetComponentKind(Type signatureType, out string kind)
+        {
+            Contracts.CheckValue(signatureType, nameof(signatureType));
+            // REVIEW: replace with a dictionary lookup.
+
+            var faceAttr = signatureType.GetTypeInfo().GetCustomAttributes(typeof(TlcModule.ComponentKindAttribute), false).FirstOrDefault()
+                    as TlcModule.ComponentKindAttribute;
+            kind = faceAttr == null ? null : faceAttr.Kind;
+            return faceAttr != null;
+        }
+
+        public bool TryGetComponentShortName(Type type, out string name)
+        {
+            ComponentInfo component;
+            if (!TryFindComponent(type, out component))
+            {
+                name = null;
+                return false;
+            }
+
+            name = component.Aliases != null && component.Aliases.Length > 0 ? component.Aliases[0] : component.Name;
+            return true;
+        }
+
+        /// <summary>
+        /// The valid names for the components and entry points must consist of letters, digits, underscores and dots,
+        /// and begin with a letter or digit.
+        /// </summary>
+        private static readonly Regex _nameRegex = new Regex(@"^\w[_\.\w]*$", RegexOptions.Compiled);
+        private static bool IsValidName(string name)
+        {
+            Contracts.AssertValueOrNull(name);
+            if (string.IsNullOrWhiteSpace(name))
+                return false;
+            return _nameRegex.IsMatch(name);
         }
 
         /// <summary>
