@@ -20,6 +20,9 @@ using Microsoft.ML.Runtime.Model;
 [assembly: LoadableClass(typeof(RankingPredictionTransformer<IPredictorProducing<float>>), typeof(RankingPredictionTransformer), null, typeof(SignatureLoadModel),
     "", RankingPredictionTransformer.LoaderSignature)]
 
+[assembly: LoadableClass(typeof(AnomalyPredictionTransformer<IPredictorProducing<float>>), typeof(AnomalyPredictionTransformer), null, typeof(SignatureLoadModel),
+    "", AnomalyPredictionTransformer.LoaderSignature)]
+
 namespace Microsoft.ML.Runtime.Data
 {
 
@@ -27,7 +30,9 @@ namespace Microsoft.ML.Runtime.Data
     /// Base class for transformers with no feature column, or more than one feature columns.
     /// </summary>
     /// <typeparam name="TModel"></typeparam>
-    public abstract class PredictionTransformerBase<TModel> : IPredictionTransformer<TModel>
+    /// <typeparam name="TScorer">The Scorer used by this <see cref="IPredictionTransformer{TModel}"/></typeparam>
+    public abstract class PredictionTransformerBase<TModel, TScorer> : IPredictionTransformer<TModel>
+        where TScorer : RowToRowScorerBase
         where TModel : class, IPredictor
     {
         /// <summary>
@@ -41,21 +46,23 @@ namespace Microsoft.ML.Runtime.Data
         protected ISchemaBindableMapper BindableMapper;
         protected ISchema TrainSchema;
 
-        public abstract bool IsRowToRowMapper { get; }
+        public bool IsRowToRowMapper => true;
+
+        protected abstract TScorer Scorer { get; set; }
 
         protected PredictionTransformerBase(IHost host, TModel model, ISchema trainSchema)
         {
             Contracts.CheckValue(host, nameof(host));
-
             Host = host;
-            Host.CheckValue(trainSchema, nameof(trainSchema));
 
+            Host.CheckValue(trainSchema, nameof(trainSchema));
             Model = model;
+
+            Host.CheckValue(trainSchema, nameof(trainSchema));
             TrainSchema = trainSchema;
         }
 
         protected PredictionTransformerBase(IHost host, ModelLoadContext ctx)
-
         {
             Host = host;
 
@@ -91,9 +98,23 @@ namespace Microsoft.ML.Runtime.Data
         /// </summary>
         /// <param name="input">The input data.</param>
         /// <returns>The transformed <see cref="IDataView"/></returns>
-        public abstract IDataView Transform(IDataView input);
 
-        public abstract IRowToRowMapper GetRowToRowMapper(ISchema inputSchema);
+        public IDataView Transform(IDataView input)
+        {
+            Host.CheckValue(input, nameof(input));
+            return Scorer.ApplyToData(Host, input);
+        }
+
+        /// <summary>
+        /// Gets a IRowToRowMapper instance.
+        /// </summary>
+        /// <param name="inputSchema"></param>
+        /// <returns></returns>
+        public IRowToRowMapper GetRowToRowMapper(ISchema inputSchema)
+        {
+            Host.CheckValue(inputSchema, nameof(inputSchema));
+            return (IRowToRowMapper)Scorer.ApplyToData(Host, new EmptyDataView(Host, inputSchema));
+        }
 
         protected void SaveModel(ModelSaveContext ctx)
         {
@@ -118,8 +139,10 @@ namespace Microsoft.ML.Runtime.Data
     /// Those are all the transformers that work with one feature column.
     /// </summary>
     /// <typeparam name="TModel">The model used to transform the data.</typeparam>
-    public abstract class SingleFeaturePredictionTransformerBase<TModel> : PredictionTransformerBase<TModel>, ISingleFeaturePredictionTransformer<TModel>, ICanSaveModel
+    /// <typeparam name="TScorer">The scorer used on this PredictionTransformer.</typeparam>
+    public abstract class SingleFeaturePredictionTransformerBase<TModel, TScorer> : PredictionTransformerBase<TModel, TScorer>, ISingleFeaturePredictionTransformer<TModel>, ICanSaveModel
         where TModel : class, IPredictor
+        where TScorer: RowToRowScorerBase
     {
         /// <summary>
         /// The name of the feature column used by the prediction transformer.
@@ -131,8 +154,17 @@ namespace Microsoft.ML.Runtime.Data
         /// </summary>
         public ColumnType FeatureColumnType { get; }
 
+        protected override TScorer Scorer { get; set; }
+
+        /// <summary>
+        /// Initializes a new reference of <see cref="SingleFeaturePredictionTransformerBase{TModel, TScorer}"/>.
+        /// </summary>
+        /// <param name="host">The local instance of <see cref="IHost"/>.</param>
+        /// <param name="model">The model used for scoring.</param>
+        /// <param name="trainSchema">The schema of the training data.</param>
+        /// <param name="featureColumn">The feature column name.</param>
         public SingleFeaturePredictionTransformerBase(IHost host, TModel model, ISchema trainSchema, string featureColumn)
-            :base(host, model, trainSchema)
+            : base(host, model, trainSchema)
         {
             FeatureColumn = featureColumn;
 
@@ -148,7 +180,7 @@ namespace Microsoft.ML.Runtime.Data
         }
 
         internal SingleFeaturePredictionTransformerBase(IHost host, ModelLoadContext ctx)
-            :base(host, ctx)
+            : base(host, ctx)
         {
             FeatureColumn = ctx.LoadStringOrNull();
 
@@ -166,7 +198,7 @@ namespace Microsoft.ML.Runtime.Data
         {
             Host.CheckValue(inputSchema, nameof(inputSchema));
 
-            if(FeatureColumn != null)
+            if (FeatureColumn != null)
             {
                 if (!inputSchema.TryGetColumnIndex(FeatureColumn, out int col))
                     throw Host.ExceptSchemaMismatch(nameof(inputSchema), RoleMappedSchema.ColumnRole.Feature.Value, FeatureColumn, FeatureColumnType.ToString(), null);
@@ -189,17 +221,89 @@ namespace Microsoft.ML.Runtime.Data
             SaveModel(ctx);
             ctx.SaveStringOrNull(FeatureColumn);
         }
+
+        protected virtual GenericScorer GetGenericScorer()
+        {
+            var schema = new RoleMappedSchema(TrainSchema, null, FeatureColumn);
+            return new GenericScorer(Host, new GenericScorer.Arguments(), new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
+        }
+    }
+
+    /// <summary>
+    /// Base class for the <see cref="ISingleFeaturePredictionTransformer{TModel}"/> working on anomaly detection tasks.
+    /// </summary>
+    /// <typeparam name="TModel">An implementation of the <see cref="IPredictorProducing{TResult}"/></typeparam>
+    public sealed class AnomalyPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel, BinaryClassifierScorer>
+        where TModel : class, IPredictorProducing<float>
+    {
+        public readonly string ThresholdColumn;
+        public readonly float Threshold;
+
+        public AnomalyPredictionTransformer(IHostEnvironment env, TModel model, ISchema inputSchema, string featureColumn,
+            float threshold = 0f, string thresholdColumn = DefaultColumnNames.Score)
+            : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(BinaryPredictionTransformer<TModel>)), model, inputSchema, featureColumn)
+        {
+            Host.CheckNonEmpty(thresholdColumn, nameof(thresholdColumn));
+            Threshold = threshold;
+            ThresholdColumn = thresholdColumn;
+
+            SetScorer();
+        }
+
+        public AnomalyPredictionTransformer(IHostEnvironment env, ModelLoadContext ctx)
+            : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(BinaryPredictionTransformer<TModel>)), ctx)
+        {
+            // *** Binary format ***
+            // <base info>
+            // float: scorer threshold
+            // id of string: scorer threshold column
+
+            Threshold = ctx.Reader.ReadSingle();
+            ThresholdColumn = ctx.LoadString();
+            SetScorer();
+        }
+
+        private void SetScorer()
+        {
+            var schema = new RoleMappedSchema(TrainSchema, null, FeatureColumn);
+            var args = new BinaryClassifierScorer.Arguments { Threshold = Threshold, ThresholdColumn = ThresholdColumn };
+            Scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
+        }
+
+        protected override void SaveCore(ModelSaveContext ctx)
+        {
+            Contracts.AssertValue(ctx);
+            ctx.SetVersionInfo(GetVersionInfo());
+
+            // *** Binary format ***
+            // <base info>
+            // float: scorer threshold
+            // id of string: scorer threshold column
+            base.SaveCore(ctx);
+
+            ctx.Writer.Write(Threshold);
+            ctx.SaveString(ThresholdColumn);
+        }
+
+        private static VersionInfo GetVersionInfo()
+        {
+            return new VersionInfo(
+                modelSignature: "ANOMPRED",
+                verWrittenCur: 0x00010001, // Initial
+                verReadableCur: 0x00010001,
+                verWeCanReadBack: 0x00010001,
+                loaderSignature: AnomalyPredictionTransformer.LoaderSignature,
+                loaderAssemblyName: typeof(AnomalyPredictionTransformer<>).Assembly.FullName);
+        }
     }
 
     /// <summary>
     /// Base class for the <see cref="ISingleFeaturePredictionTransformer{TModel}"/> working on binary classification tasks.
     /// </summary>
     /// <typeparam name="TModel">An implementation of the <see cref="IPredictorProducing{TResult}"/></typeparam>
-    public sealed class BinaryPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel>
+    public sealed class BinaryPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel, BinaryClassifierScorer>
         where TModel : class, IPredictorProducing<float>
     {
-        private readonly BinaryClassifierScorer _scorer;
-
         public readonly string ThresholdColumn;
         public readonly float Threshold;
 
@@ -208,12 +312,10 @@ namespace Microsoft.ML.Runtime.Data
             : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(BinaryPredictionTransformer<TModel>)), model, inputSchema, featureColumn)
         {
             Host.CheckNonEmpty(thresholdColumn, nameof(thresholdColumn));
-            var schema = new RoleMappedSchema(inputSchema, null, featureColumn);
             Threshold = threshold;
             ThresholdColumn = thresholdColumn;
 
-            var args = new BinaryClassifierScorer.Arguments { Threshold = Threshold, ThresholdColumn = ThresholdColumn };
-            _scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, inputSchema), BindableMapper.Bind(Host, schema), schema);
+            SetScorer();
         }
 
         public BinaryPredictionTransformer(IHostEnvironment env, ModelLoadContext ctx)
@@ -226,24 +328,14 @@ namespace Microsoft.ML.Runtime.Data
 
             Threshold = ctx.Reader.ReadSingle();
             ThresholdColumn = ctx.LoadString();
+            SetScorer();
+        }
 
+        private void SetScorer()
+        {
             var schema = new RoleMappedSchema(TrainSchema, null, FeatureColumn);
             var args = new BinaryClassifierScorer.Arguments { Threshold = Threshold, ThresholdColumn = ThresholdColumn };
-            _scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
-        }
-
-        public override IDataView Transform(IDataView input)
-        {
-            Host.CheckValue(input, nameof(input));
-            return _scorer.ApplyToData(Host, input);
-        }
-
-        public override bool IsRowToRowMapper => true;
-
-        public override IRowToRowMapper GetRowToRowMapper(ISchema inputSchema)
-        {
-            Host.CheckValue(inputSchema, nameof(inputSchema));
-            return (IRowToRowMapper)_scorer.ApplyToData(Host, new EmptyDataView(Host, inputSchema));
+            Scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
         }
 
         protected override void SaveCore(ModelSaveContext ctx)
@@ -277,10 +369,9 @@ namespace Microsoft.ML.Runtime.Data
     /// Base class for the <see cref="ISingleFeaturePredictionTransformer{TModel}"/> working on multi-class classification tasks.
     /// </summary>
     /// <typeparam name="TModel">An implementation of the <see cref="IPredictorProducing{TResult}"/></typeparam>
-    public sealed class MulticlassPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel>
+    public sealed class MulticlassPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel, MultiClassClassifierScorer>
         where TModel : class, IPredictorProducing<VBuffer<float>>
     {
-        private readonly MultiClassClassifierScorer _scorer;
         private readonly string _trainLabelColumn;
 
         public MulticlassPredictionTransformer(IHostEnvironment env, TModel model, ISchema inputSchema, string featureColumn, string labelColumn)
@@ -289,9 +380,7 @@ namespace Microsoft.ML.Runtime.Data
             Host.CheckValueOrNull(labelColumn);
 
             _trainLabelColumn = labelColumn;
-            var schema = new RoleMappedSchema(inputSchema, labelColumn, featureColumn);
-            var args = new MultiClassClassifierScorer.Arguments();
-            _scorer = new MultiClassClassifierScorer(Host, args, new EmptyDataView(Host, inputSchema), BindableMapper.Bind(Host, schema), schema);
+            SetScorer();
         }
 
         public MulticlassPredictionTransformer(IHostEnvironment env, ModelLoadContext ctx)
@@ -302,24 +391,14 @@ namespace Microsoft.ML.Runtime.Data
             // id of string: train label column
 
             _trainLabelColumn = ctx.LoadStringOrNull();
+            SetScorer();
+        }
 
+        private void SetScorer()
+        {
             var schema = new RoleMappedSchema(TrainSchema, _trainLabelColumn, FeatureColumn);
             var args = new MultiClassClassifierScorer.Arguments();
-            _scorer = new MultiClassClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
-        }
-
-        public override IDataView Transform(IDataView input)
-        {
-            Host.CheckValue(input, nameof(input));
-            return _scorer.ApplyToData(Host, input);
-        }
-
-        public override bool IsRowToRowMapper => true;
-
-        public override IRowToRowMapper GetRowToRowMapper(ISchema inputSchema)
-        {
-            Host.CheckValue(inputSchema, nameof(inputSchema));
-            return (IRowToRowMapper)_scorer.ApplyToData(Host, new EmptyDataView(Host, inputSchema));
+            Scorer = new MultiClassClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
         }
 
         protected override void SaveCore(ModelSaveContext ctx)
@@ -351,37 +430,19 @@ namespace Microsoft.ML.Runtime.Data
     /// Base class for the <see cref="ISingleFeaturePredictionTransformer{TModel}"/> working on regression tasks.
     /// </summary>
     /// <typeparam name="TModel">An implementation of the <see cref="IPredictorProducing{TResult}"/></typeparam>
-    public sealed class RegressionPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel>
+    public sealed class RegressionPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel, GenericScorer>
         where TModel : class, IPredictorProducing<float>
     {
-        private readonly GenericScorer _scorer;
-
         public RegressionPredictionTransformer(IHostEnvironment env, TModel model, ISchema inputSchema, string featureColumn)
             : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(RegressionPredictionTransformer<TModel>)), model, inputSchema, featureColumn)
         {
-            var schema = new RoleMappedSchema(inputSchema, null, featureColumn);
-            _scorer = new GenericScorer(Host, new GenericScorer.Arguments(), new EmptyDataView(Host, inputSchema), BindableMapper.Bind(Host, schema), schema);
+            Scorer = GetGenericScorer();
         }
 
         internal RegressionPredictionTransformer(IHostEnvironment env, ModelLoadContext ctx)
             : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(RegressionPredictionTransformer<TModel>)), ctx)
         {
-            var schema = new RoleMappedSchema(TrainSchema, null, FeatureColumn);
-            _scorer = new GenericScorer(Host, new GenericScorer.Arguments(), new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
-        }
-
-        public override IDataView Transform(IDataView input)
-        {
-            Host.CheckValue(input, nameof(input));
-            return _scorer.ApplyToData(Host, input);
-        }
-
-        public override bool IsRowToRowMapper => true;
-
-        public override IRowToRowMapper GetRowToRowMapper(ISchema inputSchema)
-        {
-            Host.CheckValue(inputSchema, nameof(inputSchema));
-            return (IRowToRowMapper)_scorer.ApplyToData(Host, new EmptyDataView(Host, inputSchema));
+            Scorer = GetGenericScorer();
         }
 
         protected override void SaveCore(ModelSaveContext ctx)
@@ -397,7 +458,7 @@ namespace Microsoft.ML.Runtime.Data
         private static VersionInfo GetVersionInfo()
         {
             return new VersionInfo(
-                modelSignature: "MC  PRED",
+                modelSignature: "REG PRED",
                 verWrittenCur: 0x00010001, // Initial
                 verReadableCur: 0x00010001,
                 verWeCanReadBack: 0x00010001,
@@ -406,37 +467,23 @@ namespace Microsoft.ML.Runtime.Data
         }
     }
 
-    public sealed class RankingPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel>
+    /// <summary>
+    /// Base class for the <see cref="ISingleFeaturePredictionTransformer{TModel}"/> working on ranking tasks.
+    /// </summary>
+    /// <typeparam name="TModel">An implementation of the <see cref="IPredictorProducing{TResult}"/></typeparam>
+    public sealed class RankingPredictionTransformer<TModel> : SingleFeaturePredictionTransformerBase<TModel, GenericScorer>
     where TModel : class, IPredictorProducing<float>
     {
-        private readonly GenericScorer _scorer;
-
         public RankingPredictionTransformer(IHostEnvironment env, TModel model, ISchema inputSchema, string featureColumn)
             : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(RankingPredictionTransformer<TModel>)), model, inputSchema, featureColumn)
         {
-            var schema = new RoleMappedSchema(inputSchema, null, featureColumn);
-            _scorer = new GenericScorer(Host, new GenericScorer.Arguments(), new EmptyDataView(Host, inputSchema), BindableMapper.Bind(Host, schema), schema);
+            Scorer = GetGenericScorer();
         }
 
         internal RankingPredictionTransformer(IHostEnvironment env, ModelLoadContext ctx)
             : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(RankingPredictionTransformer<TModel>)), ctx)
         {
-            var schema = new RoleMappedSchema(TrainSchema, null, FeatureColumn);
-            _scorer = new GenericScorer(Host, new GenericScorer.Arguments(), new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
-        }
-
-        public override IDataView Transform(IDataView input)
-        {
-            Host.CheckValue(input, nameof(input));
-            return _scorer.ApplyToData(Host, input);
-        }
-
-        public override bool IsRowToRowMapper => true;
-
-        public override IRowToRowMapper GetRowToRowMapper(ISchema inputSchema)
-        {
-            Host.CheckValue(inputSchema, nameof(inputSchema));
-            return (IRowToRowMapper)_scorer.ApplyToData(Host, new EmptyDataView(Host, inputSchema));
+            Scorer = GetGenericScorer();
         }
 
         protected override void SaveCore(ModelSaveContext ctx)
@@ -452,7 +499,7 @@ namespace Microsoft.ML.Runtime.Data
         private static VersionInfo GetVersionInfo()
         {
             return new VersionInfo(
-                modelSignature: "MC  RANK",
+                modelSignature: "RANK PRED",
                 verWrittenCur: 0x00010001, // Initial
                 verReadableCur: 0x00010001,
                 verWeCanReadBack: 0x00010001,
@@ -491,5 +538,13 @@ namespace Microsoft.ML.Runtime.Data
 
         public static RankingPredictionTransformer<IPredictorProducing<float>> Create(IHostEnvironment env, ModelLoadContext ctx)
             => new RankingPredictionTransformer<IPredictorProducing<float>>(env, ctx);
+    }
+
+    internal static class AnomalyPredictionTransformer
+    {
+        public const string LoaderSignature = "AnomalyPredXfer";
+
+        public static AnomalyPredictionTransformer<IPredictorProducing<float>> Create(IHostEnvironment env, ModelLoadContext ctx)
+            => new AnomalyPredictionTransformer<IPredictorProducing<float>>(env, ctx);
     }
 }
