@@ -59,7 +59,7 @@ namespace Microsoft.ML.Runtime.EntryPoints.JsonUtils
                 throw _host.Except(ex, "Failed to parse experiment graph: {0}", ex.Message);
             }
 
-            var catalog = ModuleCatalog.CreateInstance(_host);
+            var catalog = _host.ComponentCatalog;
             var runner = new GraphRunner(_host, catalog, graph[FieldNames.Nodes] as JArray);
 
             // Setting inputs.
