@@ -29,7 +29,7 @@ namespace Microsoft.ML.Runtime
         }
 
         private readonly Runtime.IHostEnvironment _env;
-        private readonly ModuleCatalog _catalog;
+        private readonly ComponentCatalog _catalog;
         private readonly List<string> _jsonNodes;
         private readonly JsonSerializer _serializer;
         private readonly SerializationHelper _helper;
@@ -39,7 +39,7 @@ namespace Microsoft.ML.Runtime
             _env = env;
             AssemblyLoadingUtils.RegisterCurrentLoadedAssemblies(_env);
 
-            _catalog = ModuleCatalog.CreateInstance(_env);
+            _catalog = _env.ComponentCatalog;
             _jsonNodes = new List<string>();
             _serializer = new JsonSerializer();
             _serializer.Converters.Add(new StringEnumConverter());
