@@ -1584,7 +1584,7 @@ namespace Microsoft.ML.Runtime.PipelineInference
             h.Check(args.EstimatedSampleFraction > 0);
 
             //available transforms in this environment
-            var availableTransforms = ModuleCatalog.CreateInstance(env).AllEntryPoints()
+            var availableTransforms = env.ComponentCatalog.AllEntryPoints()
                 .Where(x => x.InputKinds?.FirstOrDefault(i => i == typeof(CommonInputs.ITransformInput)) != null
                     && x.InputKinds?.Any(i => i == typeof(CommonInputs.ICalibratorInput)) != true);
             var dataSample = data.Take(MaxRowsToRead);
