@@ -190,7 +190,7 @@ namespace Microsoft.ML.Runtime.FactorizationMachine
         }
     }
 
-    public sealed class FieldAwareFactorizationMachinePredictionTransformer : PredictionTransformerBase<FieldAwareFactorizationMachinePredictor>, ICanSaveModel
+    public sealed class FieldAwareFactorizationMachinePredictionTransformer : PredictionTransformerBase<FieldAwareFactorizationMachinePredictor, BinaryClassifierScorer>, ICanSaveModel
     {
         public const string LoaderSignature = "FAFMPredXfer";
 
@@ -204,7 +204,7 @@ namespace Microsoft.ML.Runtime.FactorizationMachine
         /// </summary>
         public ColumnType[] FeatureColumnTypes { get; }
 
-        private readonly BinaryClassifierScorer _scorer;
+        protected override BinaryClassifierScorer Scorer { get; set; }
 
         private readonly string _thresholdColumn;
         private readonly float _threshold;
@@ -236,7 +236,7 @@ namespace Microsoft.ML.Runtime.FactorizationMachine
 
             var schema = GetSchema();
             var args = new BinaryClassifierScorer.Arguments { Threshold = _threshold, ThresholdColumn = _thresholdColumn };
-            _scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, trainSchema), BindableMapper.Bind(Host, schema), schema);
+            Scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, trainSchema), BindableMapper.Bind(Host, schema), schema);
         }
 
         public FieldAwareFactorizationMachinePredictionTransformer(IHostEnvironment host, ModelLoadContext ctx)
@@ -269,11 +269,11 @@ namespace Microsoft.ML.Runtime.FactorizationMachine
 
             var schema = GetSchema();
             var args = new BinaryClassifierScorer.Arguments { Threshold = _threshold, ThresholdColumn = _thresholdColumn };
-            _scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
+            Scorer = new BinaryClassifierScorer(Host, args, new EmptyDataView(Host, TrainSchema), BindableMapper.Bind(Host, schema), schema);
         }
 
         /// <summary>
-        /// Gets the <see cref="ISchema"/> result after applying <see cref="Transform(IDataView)"/>.
+        /// Gets the <see cref="ISchema"/> result after transformation.
         /// </summary>
         /// <param name="inputSchema">The <see cref="ISchema"/> of the input data.</param>
         /// <returns>The post transformation <see cref="ISchema"/>.</returns>
@@ -290,25 +290,6 @@ namespace Microsoft.ML.Runtime.FactorizationMachine
             }
 
             return Transform(new EmptyDataView(Host, inputSchema)).Schema;
-        }
-
-        /// <summary>
-        /// Applies the transformer to the <paramref name="input"/>, scoring it through the <see cref="BinaryClassifierScorer"/>.
-        /// </summary>
-        /// <param name="input">The data to be scored with the <see cref="FieldAwareFactorizationMachinePredictor"/>.</param>
-        /// <returns>The scored <see cref="IDataView"/>.</returns>
-        public override IDataView Transform(IDataView input)
-        {
-            Host.CheckValue(input, nameof(input));
-            return _scorer.ApplyToData(Host, input);
-        }
-
-        public override bool IsRowToRowMapper => true;
-
-        public override IRowToRowMapper GetRowToRowMapper(ISchema inputSchema)
-        {
-            Host.CheckValue(inputSchema, nameof(inputSchema));
-            return (IRowToRowMapper)_scorer.ApplyToData(Host, new EmptyDataView(Host, inputSchema));
         }
 
         /// <summary>
