@@ -249,7 +249,7 @@ namespace Microsoft.ML.Runtime.RunTests
             var epListFile = buildPrefix + "_ep-list.tsv";
 
             var entryPointsSubDir = Path.Combine("..", "Common", "EntryPoints");
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var epListPath = GetBaselinePath(entryPointsSubDir, epListFile);
             DeleteOutputPath(epListPath);
 
@@ -277,7 +277,7 @@ namespace Microsoft.ML.Runtime.RunTests
             var epListFile = buildPrefix + "_ep-list.tsv";
 
             var entryPointsSubDir = Path.Combine("..", "Common", "EntryPoints");
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var path = DeleteOutputPath(entryPointsSubDir, epListFile);
 
             File.WriteAllLines(path, epListContents);
@@ -307,7 +307,7 @@ namespace Microsoft.ML.Runtime.RunTests
             Env.ComponentCatalog.RegisterAssembly(typeof(AutoInference).Assembly);
             Env.ComponentCatalog.RegisterAssembly(typeof(SaveOnnxCommand).Assembly);
 
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
 
             var regex = new Regex(@"\r\n?|\n", RegexOptions.Compiled);
             var epListContents = catalog.AllEntryPoints()
@@ -337,7 +337,7 @@ namespace Microsoft.ML.Runtime.RunTests
         [Fact]
         public void EntryPointInputBuilderOptionals()
         {
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
 
             InputBuilder ib1 = new InputBuilder(Env, typeof(LogisticRegression.Arguments), catalog);
             // Ensure that InputBuilder unwraps the Optional<string> correctly.
@@ -558,7 +558,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -662,7 +662,7 @@ namespace Microsoft.ML.Runtime.RunTests
         //        }";
 
         //    JObject graph = JObject.Parse(inputGraph);
-        //    var catalog = ModuleCatalog.CreateInstance(Env);
+        //    var catalog = Env.ComponentCatalog;
         //    var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
         //    var dv1_data = new[]
@@ -1215,7 +1215,7 @@ namespace Microsoft.ML.Runtime.RunTests
                     Column = new[] { new ConcatTransform.Column() { Name = "Features", Source = new[] { "Features1", "Features2" } } }
                 }, data);
 
-                var mlr = new MulticlassLogisticRegression(Env, new MulticlassLogisticRegression.Arguments());
+                var mlr = new MulticlassLogisticRegression(Env, "Features", "Label");
                 var rmd = new RoleMappedData(data, "Label", "Features");
 
                 predictorModels[i] = new PredictorModel(Env, rmd, data, mlr.Train(rmd));
@@ -2307,8 +2307,8 @@ namespace Microsoft.ML.Runtime.RunTests
         [Fact]
         public void TestInputBuilderBasicArgs()
         {
-            var catalog = ModuleCatalog.CreateInstance(Env);
-            bool success = catalog.TryFindEntryPoint("Transforms.MinMaxNormalizer", out ModuleCatalog.EntryPointInfo info);
+            var catalog = Env.ComponentCatalog;
+            bool success = catalog.TryFindEntryPoint("Transforms.MinMaxNormalizer", out ComponentCatalog.EntryPointInfo info);
             Assert.True(success);
             var inputBuilder = new InputBuilder(Env, info.InputType, catalog);
 
@@ -2364,8 +2364,8 @@ namespace Microsoft.ML.Runtime.RunTests
         [Fact]
         public void TestInputBuilderComponentFactories()
         {
-            var catalog = ModuleCatalog.CreateInstance(Env);
-            bool success = catalog.TryFindEntryPoint("Trainers.StochasticDualCoordinateAscentBinaryClassifier", out ModuleCatalog.EntryPointInfo info);
+            var catalog = Env.ComponentCatalog;
+            bool success = catalog.TryFindEntryPoint("Trainers.StochasticDualCoordinateAscentBinaryClassifier", out ComponentCatalog.EntryPointInfo info);
             Assert.True(success);
             var inputBuilder = new InputBuilder(Env, info.InputType, catalog);
 
@@ -2480,7 +2480,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -2576,7 +2576,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -2684,7 +2684,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -2789,7 +2789,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -2950,7 +2950,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -3145,7 +3145,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
 
             var dataPath = GetDataPath("breast-cancer.txt");
@@ -3268,7 +3268,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graphJson = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var graph = new EntryPointGraph(Env, catalog, graphJson[FieldNames.Nodes] as JArray);
             Assert.True(graph.Macros.All(x => x.CanStart()));
         }
@@ -3323,7 +3323,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graphJson = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var graph = new EntryPointGraph(Env, catalog, graphJson[FieldNames.Nodes] as JArray);
             // Serialize the nodes with ToJson() and then executing them to ensure serialization working correctly.
             var nodes = new JArray(graph.AllNodes.Select(node => node.ToJson()));
@@ -3383,7 +3383,7 @@ namespace Microsoft.ML.Runtime.RunTests
                   ]
                 }";
             JObject graphJson = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var graph = new EntryPointGraph(Env, catalog, graphJson[FieldNames.Nodes] as JArray);
             for (int i = 0; i < 2; i++)
             {
