@@ -80,7 +80,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
 
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
             runner.SetInput("TrainingData", datasetTrain);
@@ -218,7 +218,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graphJson = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var graph = new EntryPointGraph(Env, catalog, graphJson[FieldNames.Nodes] as JArray);
             // Test if ToJson() works properly.
             var nodes = new JArray(graph.AllNodes.Select(node => node.ToJson()));
@@ -299,7 +299,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graphJson = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graphJson[FieldNames.Nodes] as JArray);
             runner.SetInput("TrainingData", datasetTrain);
             runner.SetInput("TestingData", datasetTest);
@@ -383,7 +383,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graphJson = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
             var runner = new GraphRunner(Env, catalog, graphJson[FieldNames.Nodes] as JArray);
             runner.SetInput("TrainingData", datasetTrain);
             runner.SetInput("TestingData", datasetTest);
@@ -467,7 +467,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
 
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
             runner.SetInput("TrainingData", datasetTrain);
@@ -551,7 +551,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 }";
 
             JObject graph = JObject.Parse(inputGraph);
-            var catalog = ModuleCatalog.CreateInstance(Env);
+            var catalog = Env.ComponentCatalog;
 
             var runner = new GraphRunner(Env, catalog, graph[FieldNames.Nodes] as JArray);
             runner.SetInput("TrainingData", datasetTrain);
