@@ -55,12 +55,8 @@ namespace Microsoft.ML.Scenarios
                 loader);
 
                 // Train
-                var trainer = new FastTreeBinaryClassificationTrainer(env, DefaultColumnNames.Label, DefaultColumnNames.Features, advancedSettings: s=>
-                {
-                    s.NumLeaves = 5;
-                    s.NumTrees = 5;
-                    s.MinDocumentsInLeafs = 2;
-                });
+                var trainer = new FastTreeBinaryClassificationTrainer(env, DefaultColumnNames.Label, DefaultColumnNames.Features, 
+                    numLeaves:5, numTrees:5, minDocumentsInLeafs: 2);
 
                 var trainRoles = new RoleMappedData(trans, label: "Label", feature: "Features");
                 var pred = trainer.Train(trainRoles);
@@ -136,12 +132,7 @@ namespace Microsoft.ML.Scenarios
                     ModelKind = WordEmbeddingsTransform.PretrainedModelKind.Sswe,
                 }, text);
                 // Train
-                var trainer = new FastTreeBinaryClassificationTrainer(env, DefaultColumnNames.Label, DefaultColumnNames.Features, advancedSettings: s=> 
-                    {
-                        s.NumLeaves = 5;
-                        s.NumTrees = 5;
-                        s.MinDocumentsInLeafs = 2;
-                    });
+                var trainer = new FastTreeBinaryClassificationTrainer(env, DefaultColumnNames.Label, DefaultColumnNames.Features, numLeaves: 5, numTrees:5, minDocumentsInLeafs:2);
 
                 var trainRoles = new RoleMappedData(trans, label: "Label", feature: "Features");
                 var pred = trainer.Train(trainRoles);
