@@ -547,7 +547,7 @@ namespace Microsoft.ML.Scenarios
             });
             pipeline.Add(new Legacy.Transforms.ColumnConcatenator() { Column = new[] { new ConcatTransformColumn() { Name = "Features", Source = new[] { "Placeholder", "dense/Relu" } } } });
             pipeline.Add(new Legacy.Trainers.LogisticRegressionClassifier());
-            TensorFlowUtils.Initialize();
+
             var model = pipeline.Train<MNISTData, MNISTPrediction>();
 
             var sample1 = new MNISTData()
@@ -658,7 +658,14 @@ namespace Microsoft.ML.Scenarios
                 var imageWidth = 32;
                 var dataFile = GetDataPath("images/images.tsv");
                 var imageFolder = Path.GetDirectoryName(dataFile);
-                var data = env.CreateLoader("Text{col=ImagePath:TX:0 col=Name:TX:1}", new MultiFileSource(dataFile));
+                var data = TextLoader.Create(env, new TextLoader.Arguments()
+                {
+                    Column = new[]
+                    {
+                        new TextLoader.Column("ImagePath", DataKind.TX, 0),
+                        new TextLoader.Column("Name", DataKind.TX, 1),
+                    }
+                }, new MultiFileSource(dataFile));
                 var images = ImageLoaderTransform.Create(env, new ImageLoaderTransform.Arguments()
                 {
                     Column = new ImageLoaderTransform.Column[1]
